refactor(libtool): extract schema utilities and rename test file
Move mapFieldToSchema, generateSchemaFromProtobuf, and
buildToolDescription from processor/tool.js to schema.js so the
processor file contains only the ToolProcessor class. Re-export the
utilities from index.js. Rename test/tool.test.js to
test/processor.test.js for local relevance.
