+++ libraries/libtool/index.js
@@ -1,1 +1,6 @@
 export { ToolProcessor } from "./processor/tool.js";
+export {
+  mapFieldToSchema,
+  generateSchemaFromProtobuf,
+  buildToolDescription,
+} from "./schema.js";
